refactor: remove "Page" suffix from files in pages folder
