+++ frontend/src/App.tsx
@@ -3,15 +3,15 @@ import './App.css';
 import { Route, Routes } from 'react-router-dom';
 
 import AppLayout from './components/layout/AppLayout';
-import HomePage from './components/pages/HomePage';
-import StandingsPage from './components/pages/StandingsPage';
+import Home from './components/pages/Home';
+import Standings from './components/pages/Standings';
 
 function App() {
   return (
     <Routes>
       <Route element={<AppLayout />}>
-        <Route path="/" element={<HomePage />} />
-        <Route path="/standings" element={<StandingsPage />} />
+        <Route path="/" element={<Home />} />
+        <Route path="/standings" element={<Standings />} />
       </Route>
     </Routes>
   );
