+++ frontend/src/components/pages/Standings.tsx
